requirements.txt
django
django-cors-headers
djangorestframework
mysqlclient
channels
channels-redis
firebase-admin